edited DepartmentService.cs, DepartmentController.cs, BonusController.cs, BonusService.cs

+++ SyudentAccounting.BusinessLogic/Services/Implementations/DepartmentService.cs
@@ -2,16 +2,21 @@
 using StudentAccounting.Model;
 using StudentAccounting.BusinessLogic.Services.Contracts;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace StudentAccounting.BusinessLogic.Services.Implementations
 {
     public class DepartmentService : IDepartmentService
     {
+        private readonly ILogger<DepartmentService> _logger;
         private readonly ApplicationDatabaseContext _context;
-        public DepartmentService(ApplicationDatabaseContext context)
+        
+        public DepartmentService(ApplicationDatabaseContext context, ILogger<DepartmentService> logger)
         {
+            _logger = logger;
             _context = context;
         }
+        
         public void Create(Department department)
         {
             try
@@ -21,7 +26,7 @@ namespace StudentAccounting.BusinessLogic.Services.Implementations
             }
             catch (DbUpdateException ex)
             {
-                throw new Exception(ex.Message);
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
             }
         }
 
@@ -29,12 +34,15 @@ namespace StudentAccounting.BusinessLogic.Services.Implementations
         {
             try
             {
-                return _context.Departments.Include(x => x.Organizations).AsNoTracking().ToList();
+                var departments = _context.Departments.Include(x => x.Organizations).AsNoTracking().ToList();
+
+                return departments;
             }
             catch (Exception ex)
             {
-                throw new Exception(ex.Message);
-              
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
+
+                return new List<Department>();
             }
         }
 
@@ -42,11 +50,20 @@ namespace StudentAccounting.BusinessLogic.Services.Implementations
         {
             try
             {
-                return _context.Departments.Include(x => x.Organizations).AsNoTracking().FirstOrDefault(x => x.FullName == name);
+                var department = _context.Departments.Include(x => x.Organizations).AsNoTracking().FirstOrDefault(x => x.FullName == name);
+                
+                if (department == null)
+                {
+                    return new Department();
+                }
+                
+                return department;
             }
             catch (Exception ex)
             {
-                throw new Exception(ex.Message);
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
+
+                return new Department();
             }
         }
 
@@ -54,11 +71,20 @@ namespace StudentAccounting.BusinessLogic.Services.Implementations
         {
             try
             {
-                return _context.Departments.Include(x => x.Organizations).AsNoTracking().FirstOrDefault(x => x.Id == id);
+                var department = _context.Departments.Include(x => x.Organizations).AsNoTracking().FirstOrDefault(x => x.Id == id);
+
+                if (department == null)
+                {
+                    return new Department();
+                }
+
+                return department;
             }
             catch (Exception ex)
             {
-                throw new Exception(ex.Message);
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
+
+                return new Department();
             }
         }
 
@@ -69,9 +95,9 @@ namespace StudentAccounting.BusinessLogic.Services.Implementations
                 _context.Departments.Update(department);
                 _context.SaveChanges();
             }
-            catch (DbUpdateException ex)
+            catch (Exception ex)
             {
-                throw new Exception(ex.Message);
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
             }
         }
 
@@ -80,12 +106,18 @@ namespace StudentAccounting.BusinessLogic.Services.Implementations
             try
             {
                 var department = _context.Departments.FirstOrDefault(x => x.Id == id);
+
+                if (department == null)
+                {
+                    return;
+                }
+                
                 _context.Departments.Remove(department);
                 _context.SaveChanges();
             }
             catch (DbUpdateException ex)
             {
-                throw new Exception(ex.Message);
+                _logger.LogError($"{DateTime.Now}: {ex.Message}");
             }
         }
     }
